Fixing 'npm run fix' command to target all files | Using Apache 2 license in package.json file | Removed the theme in home.tsx b/c it is unnecessary | Removed defaults in tsconfig.json file
